codet5-base.jackson-databind.mid-5551.idx-40220.2.mutant

+++ jackson-databind/src/main/java/com/fasterxml/jackson/databind/node/JsonNodeFactory.java
@@ -271,7 +271,7 @@ public class JsonNodeFactory
      */
     @Override
     public ValueNode numberNode(Double value) {
-        return (value == null) ? nullNode() : DoubleNode.valueOf(value.doubleValue());
+        return (value == null) ? nullNode() : DoubleNode.valueOf(value.intValue());
     }
 
     /**
